[HAMA-678]: fix input splitting
git-svn-id: https://svn.apache.org/repos/asf/hama/trunk@1411326 13f79535-47bb-0310-9956-ffa450edef68

+++ graph/src/main/java/org/apache/hama/graph/GraphJobRunner.java
@@ -509,7 +509,7 @@ public final class GraphJobRunner<V extends Writable, E extends Writable, M exte
         }
       }
 
-      int steps = (int) (maxSplitSize / conf.getInt( // 20 mb
+      int steps = (int) (maxSplitSize / conf.getLong( // 20 mb
           "hama.graph.multi.step.partitioning.interval", 20000000)) + 1;
 
       for (String peerName : peer.getAllPeerNames()) {
